requirements.txt
absl-py==0.11.0
astunparse==1.6.3
cachetools==4.2.0
certifi==2020.12.5
chardet==3.0.4
cycler==0.10.0
decorator==4.4.2
gast==0.3.3
imutils==0.5.3
numpy==1.18.5
opencv-contrib-python==4.1.0.25
Pillow==8.0.1
scikit-image==0.17.2
scikit-learn==0.22.2
scipy==1.4.1
six==1.15.0
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorflow @ file:///home/pi/tensorflow-2.3.0-cp37-none-linux_armv7l.whl
tensorflow-estimator==2.3.0
